src/requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
Jinja2==2.11.3
MarkupSafe==1.1.1
neo4j==4.2.1
pydantic==1.8.1
python-dotenv==0.15.0
pytz==2021.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
